requirement.txt
numpy
streamlit
transformers
sounddevice
pinecone-client
huggingface-hub
sentence-transformers
requests
torch
pymongo